Add Most Popular movies rating view

+++ android/app/src/main/java/com/backbase/assignment/ui/home/model/MostPopularItem.kt
@@ -1,3 +1,3 @@
 package com.backbase.assignment.ui.home.model
 
-data class MostPopularItem(val id: Int, val posterPath: String, val title: String, val releaseDate: String, val rating: Double)
+data class MostPopularItem(val id: Int, val posterPath: String, val title: String, val releaseDate: String, val rating: Int)
